requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
dj-rest-auth==5.0.2
Django==4.2.8
django-autoslug==1.9.9
django-cleanup==8.1.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-image-optimizer==1.0.3
django-jazzmin==2.6.0
django-markdown==0.8.4
django-oauth-toolkit==2.3.0
django-taggit==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
idna==3.6
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jwcrypto==1.5.4
Markdown==3.5.2
numpy==1.26.4
oauthlib==3.2.2
pillow==10.2.0
psycopg-binary==3.1.18
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
python-resize-image==1.1.20
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
sqlparse==0.4.4
tinify==1.6.0
typing_extensions==4.10.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0
